Change the binding logic
Bindining are operating in the reverse order. First we need to specify the token and then the provider for that token.
For example `bind(IFoo).to(Foo)`, when we now ask for object of type IFoo we get an instance of class Foo.
This is similar to the way some other IoC work e.g. Ninject, Angular 2 DI